Added some old bodyFSI classes with old rbfmesh motion solvers based on points. In addition added a simple serial implementatino for point rbf mesh motion

+++ src/RBFMeshMotionSolver/RBFCoarsening.C
@@ -64,7 +64,7 @@ namespace rbf
         coarseningMaxPoints( coarseningMaxPoints ),
         twoPointSelection( twoPointSelection ),
         surfaceCorrection( false ),
-        ratioRadiusError( 10 ),
+        ratioRadiusError( 10.0 ),
         exportTxt( exportTxt ),
         selectedPositions(),
         nbStaticFaceCentersRemove( 0 ),
@@ -105,7 +105,7 @@ namespace rbf
         int coarseningMaxPoints,
         bool twoPointSelection,
         bool surfaceCorrection,
-        int ratioRadiusError,
+        double ratioRadiusError,
         bool exportTxt
         )
         :
